% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosihap}
\alias{cosihap}
\title{Synthetic haplotype data generated from COSI software}
\format{
An object of class \code{dgCMatrix} with 10000 rows and 604 columns.
}
\source{
\url{https://genome.cshlp.org/content/15/11/1576.abstract}
}
\usage{
data(cosihap)
}
\description{
A dataset containing 10,000 haplotypes of SNP sequences
mimicking the European population generated from the COSI software
}
\keyword{datasets}
