% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_reml.R
\name{Rinv.AtG}
\alias{Rinv.AtG}
\title{Function for applying R inverse to AtG in REML EM algorithm}
\usage{
Rinv.AtG(G, AtG, GtAAtG, tau, sigma)
}
\arguments{
\item{G}{Matrix of genotype markers (size n x L)}

\item{AtG}{AtG from pre-computation}

\item{GtAAtG}{GtAAtG from pre-computation}

\item{tau}{Variance component from G main effect}

\item{sigma}{Variance component from model noise epsilon}
}
\value{
Matrix resulting from left multiplication of Rinv with input matrix AtG
}
\description{
Function for applying R inverse to AtG in REML EM algorithm
}
