% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_reml.R
\name{Rinv.u}
\alias{Rinv.u}
\title{Function for applying R inverse to u in REML EM algorithm}
\usage{
Rinv.u(G, AtG, GtAAtG, GtAu, u, tau, sigma)
}
\arguments{
\item{G}{Matrix of genotype markers (size n x L)}

\item{AtG}{AtG from precomputation}

\item{GtAAtG}{GtAAtG from precomputation}

\item{GtAu}{GtAu from precomputation}

\item{u}{u=Aty from REML EM}

\item{tau}{Variance component from G main effect}

\item{sigma}{Variance component from model noise epsilon}
}
\value{
Vector resulting from left multiplication of Rinv with input vector u
}
\description{
Function for applying R inverse to u in REML EM algorithm
}
