% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{SEAGLE}
\alias{SEAGLE}
\title{Compute score-like test statistic and p-value for GxE test with SEAGLE algorithm}
\usage{
SEAGLE(obj.SEAGLE, init.tau, init.sigma, pv = "liu")
}
\arguments{
\item{obj.SEAGLE}{Input data prepared with prep.SEAGLE function}

\item{init.tau}{Initial estimate for tau}

\item{init.sigma}{Initial estimate for sigma}

\item{pv}{Method of obtaining p-value (Either "liu" or "davies", Default is liu)}
}
\description{
This function computes the score test statistic and corresponding
p-value for the GxE test with the SEAGLE algorithm with input data
that have been prepared with the prep.SEAGLE function
}
\examples{
#dat <- makeSimData(H=cosihap, n=5000, L=100, gammaG=1, gammaGE=0, causal=40, seed=1)
#objSEAGLE <- prep.SEAGLE(y=dat$y, X=dat$X, intercept=1, E=dat$E, G=dat$G)
#res <- SEAGLE(objSEAGLE, init.tau=0.5, init.sigma=0.5)

}
