% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{Vinv}
\alias{Vinv}
\title{Function for applying V inverse in Algorithm 1}
\usage{
Vinv(G, qrM, tau_over_sigma, sigma, RHS)
}
\arguments{
\item{G}{Matrix of genotype markers (size n x L)}

\item{qrM}{Pre-computation for LxL linear system solve}

\item{tau_over_sigma}{Tau over sigma from precomputation}

\item{sigma}{Variance component from model noise epsilon}

\item{RHS}{Matrix or vector on right-hand side of V inverse}
}
\value{
Matrix or vector resulting from left multiplication of Vinv with input RHS
}
\description{
This function applies V inverse via the Woodbury matrix identity
}
